Updated student management endpoint and minor bug fixes

+++ src/main/java/org/lucas/arbackend/dto/quiz/QuizResultResponse.java
@@ -2,8 +2,9 @@ package org.lucas.arbackend.dto.quiz;
 
 import java.io.Serial;
 import java.io.Serializable;
+import java.math.BigDecimal;
 
-public record QuizResultResponse(int score, boolean passed) implements Serializable {
+public record QuizResultResponse(BigDecimal score, boolean passed) implements Serializable {
     @Serial
     private static final long serialVersionUID = 1L;
 }
